303sejung

+++ src/main/java/nodo/erp/Hr/Controller/Dep_Controller.java
@@ -3,6 +3,7 @@ package nodo.erp.Hr.Controller;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -14,9 +15,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import nodo.erp.Hr.CustomUserDetails;
 import nodo.erp.Hr.Dto.Dep_Form;
 import nodo.erp.Hr.Entity.Department;
+import nodo.erp.Hr.Entity.Employee;
 import nodo.erp.Hr.Service.Dep_Service;
+import nodo.erp.Hr.Service.Emp_Service;
 
 @RequestMapping("/dep")
 @RequiredArgsConstructor
@@ -24,6 +28,7 @@ import nodo.erp.Hr.Service.Dep_Service;
 public class Dep_Controller {
 
 	private final Dep_Service dep_Service;
+	private final Emp_Service emp_Service;
 
 	@GetMapping("/list")
 	public String list(Model model, @RequestParam(value = "page", defaultValue = "0") int page) {
@@ -33,8 +38,17 @@ public class Dep_Controller {
 	}
 
 	@GetMapping("/create")
-	public String depCreate(Dep_Form dep_Form) {
-		return "Hr/Dep_Form";
+	public String depCreate(Dep_Form dep_Form,Authentication authentication) {
+		CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
+		Employee employee = this.emp_Service.getEmpDetail(customUserDetails.getEmpid());
+		if(employee.getId() == 1) {
+			
+			return "Hr/Dep_Form";
+		}else {
+			
+			return "redirect:/";
+			
+		}
 	}
 
 	@PostMapping("/create")
@@ -71,11 +85,16 @@ public class Dep_Controller {
 
 
 	    @GetMapping("/delete/{id}")
-	    public String questionDelete(@PathVariable("id") Integer id) {
+	    public String questionDelete(@PathVariable("id") Integer id,Authentication authentication) {
 	    	Department dep = this.dep_Service.getFindByIdDep(id);
-
+	    	CustomUserDetails customUserDetails = (CustomUserDetails) authentication.getPrincipal();
+			Employee employee = this.emp_Service.getEmpDetail(customUserDetails.getEmpid());
+			if(employee.getId() == 1) {
 	        this.dep_Service.delete(dep);
 	        return "redirect:/dep/list";
+	        }else {
+	        	return "redirect:/";
+	        }
 	    }
 	    
 }
